Refactor to enable trainers to define their own fields and flags.
We want trainers that implement the interface defined in
trainer_lib.Trainer to be able to define their own fields in the
TrainerConfig. We then want to be able to set those fields via
command-line flags. To accomplish these goals, we have each
trainer  define their own TrainerConfig inheriting from
trainer_lib.TrainerConfig. We then build the command-line flags
from the fields in all the TrainerConfigs.